refactor: move ErrorBoundary to pyro components
