+++ apps/frontend/src/components/ui/servers/ErrorBoundary.vue
